requirements.txt
attrs==23.2.0
aws-cdk-lib==2.187.0
aws-cdk.asset-awscli-v1==2.2.230
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-amplify-alpha==2.133.0a0
aws-cdk.aws-apprunner-alpha==2.187.0a0
aws-cdk.cloud-assembly-schema==41.2.0
boto3==1.17.112
botocore==1.20.112
cattrs==23.2.3
cfgv==3.4.0
constructs==10.3.0
distlib==0.3.8
filelock==3.13.3
identify==2.5.35
importlib_resources==6.4.0
jmespath==0.10.0
jsii==1.110.0
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.7.0
publication==0.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
s3transfer==0.4.2
setuptools==69.2.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.10.0
urllib3==1.26.18
virtualenv==20.25.1
